requirements.txt
# Flask web framework
Flask==3.1.0
Flask-WTF==1.2.2
WTForms==3.1.0
Flask-Mail>=0.9.1
email-validator>=2.1.0
python-dotenv==0.20.0
gunicorn==21.2.0

# Core scientific computing libraries
numpy>=1.22
pandas>=2.0.0

# Parquet support (needed for BIGG)
pyarrow>=17.0.0
plotly==6.3.1

# Deep learning and graph neural networks (CPU versions)
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.0.0

# Chemistry and molecular modeling
rdkit>=2025.03.0
biopython>=1.85

# Metabolic modeling
cobra>=0.29.0

# Data visualization
matplotlib>=3.10.0
seaborn>=0.13.0

# HTTP requests and API calls
requests>=2.32.0

# Java integration for ChemAxon
jpype1>=1.6.0

# Progress bars and utilities
tqdm>=4.67.0

# Additional dependencies
pubchempy>=1.0.4
networkx>=3.5
sqlalchemy>=2.0.0

# System utilities
psutil>=7.0.0

# Additional dGbyG dependencies
portalocker>=3.2.0

